Fixing Security issues (1/2)

+++ usr/share/octopussy/report_cancel.asp
@@ -1,16 +1,17 @@
 <%
-use Octopussy;
+if (NOT_NULL($Session->{AAT_LOGIN}))
+{
+	my $pid_dir = Octopussy::FS::Directory("running");
+	my $pid_param = $Session->{progress_running};
+	my $pid_file = $pid_dir . "octo_reporter_${pid_param}.pid";
 
-my $pid_dir = Octopussy::FS::Directory("running");
-my $pid_param = $Session->{progress_running};
-my $pid_file = $pid_dir . "octo_reporter_${pid_param}.pid";
+	$Session->{progress_current} = undef;
+	$Session->{progress_desc} = undef;
+	$Session->{progress_running} = undef;
+	$Session->{progress_total} = undef;
 
-$Session->{progress_current} = undef;
-$Session->{progress_desc} = undef;
-$Session->{progress_running} = undef;
-$Session->{progress_total} = undef;
-
-my $pid = Octopussy::PID_Value($pid_file);
-kill USR2 => $pid;
-$Response->Redirect("./reports.asp");
+	my $pid = Octopussy::PID_Value($pid_file);
+	kill USR2 => $pid;
+	$Response->Redirect("./reports.asp");
+}
 %>
